requirements.txt
Package             Version
------------------- -------
Django              3.1.7
django-ckeditor     6.0.0
djangorestframework 3.12.4
flake8              3.8.4
Pillow              8.1.0
